Quick stash before refactoring saveFilter/getFilter
Do not need secure storage for filters, only username and password. Hitting a size limit when
trying to save pairings and trades with filter

These changes do show notifications on emulator. Needed to save pairings to filter. Instead of waiting
for preview, an api request is run on save.

Did some extra work with the save filter. Was working on normalizing it to one key for each filter,
and breaking away trades and pairings. Too much work though, need to save filter ids in their
own key, and iterate when retrieving, and rebuild it. Instead, better to look at normal ionic storage,
more fitting