Notifications -- Normal Storage, no initial trade data, trade dupes check, etc.
Switched from secure storage to normal storage for filter. Secure storage only needed for passwords.
Filters hit a size limit when trying to save to secure storage

Removed initial trade data on preview. Initial trade data might be outdated, as it is generated in the refresh,
not the on save api call. Adding more to on-save will make it too slow, and code will need to be refactored,
switch from callbacks to promises for api calls.

Checked duplicate trades when adding to notifications. Do not want multiples of the same added. Used indexOf

Removed pairings=false, relic code left when testing if secure storage was not saving due to size limitations

Removed commented code for local notifications. Was causing too many errors on device.