Added lastsync to filter, not main. Needed if new filter is added, so first update runs
Refactored fetch for polling into a promise. Major issue was to pass filter index to http call. If indexes are not tracked
for each http call, only the last filter receives an http update. This is due to async. Loop runs, last index is set, and
all responses are set to that last index